auth route now holds all endpoints that do not require token authentication

+++ routes/destructure.js
@@ -10,14 +10,14 @@ module.exports = d => {
 						rows: [],
 						sum: (parseFloat(row['arvestus maht']) * parseFloat(row['Hind']))
 					}
-					console.log(parseInt(row['arvestus maht']) * parseInt(row['arvestus maht']))
+					console.log(parseFloat(row['arvestus maht']) * parseFloat(row['arvestus maht']))
 					d.veoselehed[0].rows.push(row)
 				} else if(d.matched.indexOf(row) > 0){
 					let len = d.veoselehed.length
 					for(let el of d.veoselehed){
 						if(row['Elvise VL nr'] == el.VL_nr){
-							el.sum = parseInt(el.sum) + (parseInt(row['arvestus maht']) *
-							parseInt(row['arvestus maht'])) 
+							el.sum = parseFloat(el.sum) + (parseFloat(row['arvestus maht']) *
+							parseFloat(row['arvestus maht'])) 
 							el.rows.push(row)
 							break
 						} else if(row['Elvise VL nr'] != el.VL_nr){
@@ -28,7 +28,7 @@ module.exports = d => {
 								sum: (parseFloat(row['arvestus maht']) * parseFloat(row['Hind']))
 							}
 							//console.log()
-							el.sum = parseInt(el.sum) + (parseFloat(row['arvestus maht']) * parseFloat(row['Hind']))
+							el.sum = parseFloat(el.sum) + (parseFloat(row['arvestus maht']) * parseFloat(row['Hind']))
 							d.veoselehed[len].rows.push(row)
 							break
 						}
